# openapi.yaml
openapi: 3.0.0
info:
  title: StudySync API
  version: 1.1.0
  description: A collaborative study management tool API for tracking tasks, deadlines, activity, and study materials
  contact:
    name: StudySync Team
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.studysync.example.com
    description: Production server
...
paths:
  /api/members:
    get:
      tags:
        - Members
      summary: Get all team members with group information
      description: Retrieves a list of all team members including their reliability scores and group assignments
      operationId: getMembers
      responses:
        '200':
          description: Successfully retrieved list of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
              example:
                - id: anna
                  name: Anna
                  initials: A
                  reliability: 95
                  group_id: g1
                  group_name: Study Group Alpha
        '500':
          description: Server error - unable to fetch members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Failed to fetch members
    post:
      tags:
        - Members
      summary: Create a new member
      operationId: createMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - name
                - initials
              properties:
                id:
                  type: string
                  example: dan
                name:
                  type: string
                  example: Dan
                initials:
                  type: string
                  example: D
                reliability:
                  type: integer
                  example: 100
                  minimum: 0
                  maximum: 100
                group_id:
                  type: string
                  nullable: true
                  example: g1
      responses:
        '201':
          description: Member created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          description: Missing or invalid fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error - unable to create member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/members/{id}:
    put:
      tags:
        - Members
      summary: Update an existing member
      operationId: updateMember
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - initials
              properties:
                name:
                  type: string
                initials:
                  type: string
                reliability:
                  type: integer
                  minimum: 0
                  maximum: 100
                group_id:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Member updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          description: Missing or invalid fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error - unable to update member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Members
      summary: Delete a member
      operationId: deleteMember
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Member deleted
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error - unable to delete member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Member:
      type: object
      description: A team member with reliability tracking
      required:
        - id
        - name
        - initials
        - reliability
      properties:
        id:
          type: string
          example: anna
        name:
          type: string
          example: Anna
        initials:
          type: string
          example: A
          minLength: 1
          maxLength: 5
        reliability:
          type: integer
          example: 95
          minimum: 0
          maximum: 100
        group_id:
          type: string
          example: g1
          nullable: true
        group_name:
          type: string
          example: Study Group Alpha
          nullable: true
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Failed to fetch members